refactor(build): migrate dependency management to uv and hatchling
- Migrated project's dependency management from PDM to `uv`.
- Switched the build backend from `pdm-backend` to `hatchling.build` in all `pyproject.toml` files.
- 【Configuration】Updated `pyproject.toml` files to:
  - Remove `[tool.pdm]` and `[tool.pdm.dev-dependencies]` sections.
  - Add `[tool.uv.sources]` to define editable local dependencies.
  - Explicitly list local dependencies in the `dependencies` section for applications.
  - Consolidated `authors` field to a single line where applicable.
- 【New File】Introduced `app/trading/uv.lock` for locked dependencies.
- 【Developer Experience】Updated `.vscode/settings.json` to include `python.analysis.extraPaths` for improved module resolution in VS Code.
- 【Tooling】Upgraded `ruff` to version `0.14.10` in the root `pyproject.toml` and adjusted its ignore list.
