+++ app/trading/trading/transaction/__init__.py
@@ -1,10 +1,11 @@
 from trading.transaction.base import TransactionSender
 from trading.transaction.builders.base import TransactionBuilder
-from trading.transaction.factory import TradingService
+from trading.transaction.factory import AggregateTransactionBuilder, TradingService
 from trading.transaction.protocol import TradingRoute
 from trading.transaction.sender import DefaultTransactionSender, JitoTransactionSender
 
 __all__ = [
+    "AggregateTransactionBuilder",
     "DefaultTransactionSender",
     "JitoTransactionSender",
     "TradingRoute",
