+++ app/wallet-tracker/pyproject.toml
@@ -4,26 +4,21 @@ version = "0.2.1"
 description = "Wallet tracking service for Solana Trading Bot"
 authors = [{ name = "mkdir700", email = "mkdir700@gmail.com" }]
 dependencies = [
+    "solbot-cache",
+    "solbot-db",
+    "solbot-services",
+    "solbot-common",
+    "solbot-yellowstone-grpc",
 ]
 requires-python = "==3.10.*"
 license = { text = "Apache-2.0" }
 
-[build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
+[dependency-groups]
+dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]
 
-[tool.pdm]
-distribution = false
-
-[tool.pdm.dev-dependencies]
-local = [
-    "solbot-cache @ file:///${PROJECT_ROOT}/libs/cache",
-    "solbot-db @ file:///${PROJECT_ROOT}/libs/db",
-    "solbot-services @ file:///${PROJECT_ROOT}/libs/services",
-    "solbot-common @ file:///${PROJECT_ROOT}/libs/common",
-    "solbot-yellowstone-grpc @ file:///${PROJECT_ROOT}/libs/yellowstone_grpc",
-]
-dev = [
-    "pytest>=8.0.0",
-    "pytest-asyncio>=0.24.0",
-]
+[tool.uv.sources]
+solbot-cache = { path = "../../libs/cache", editable = true }
+solbot-db = { path = "../../libs/db", editable = true }
+solbot-services = { path = "../../libs/services", editable = true }
+solbot-common = { path = "../../libs/common", editable = true }
+solbot-yellowstone-grpc = { path = "../../libs/yellowstone_grpc", editable = true }
