+++ libs/db/pyproject.toml
@@ -12,5 +12,8 @@ requires-python = "==3.10.*"
 license = { text = "Apache-2.0" }
 
 [build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.uv.sources]
+solbot-common = { path = "../common", editable = true }
