+++ libs/services/pyproject.toml
@@ -15,7 +15,6 @@ dependencies = [
     "solders>=0.25.0",
     "base58>=2.1.1",
     "tomli>=2.1.0",
-    "sqlmodel>=0.0.22",
     "pydantic-settings>=2.6.1",
     "pydantic-settings-toml>=0.2.0",
     "orjson>=3.10.11",
@@ -33,7 +32,6 @@ dependencies = [
     "aiogram>=3.15.0",
     "aiohttp<3.11,>=3.9.0",
     "jinja2>=3.1.4",
-    "sqlalchemy[asyncio]>=2.0.36",
     "aiomysql>=0.2.0",
     "jupiter-python-sdk>=0.0.2.0",
     "aiocache[redis]>=0.12.3",
@@ -42,5 +40,10 @@ requires-python = "==3.10.*"
 license = { text = "Apache-2.0" }
 
 [build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.uv.sources]
+solbot-common = { path = "../common", editable = true }
+solbot-db = { path = "../db", editable = true }
+solbot-cache = { path = "../cache", editable = true }
